Extract log config package
