+++ omniwitness/public_witness_network.go
@@ -20,6 +20,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/transparency-dev/witness/config"
 	"k8s.io/klog/v2"
 )
 
@@ -43,12 +44,12 @@ func provisionFromPublicConfig(ctx context.Context, httpClient *http.Client, url
 		}
 		klog.V(1).Infof("Provisioning from public witness network...")
 		for _, u := range urls {
-			opts := PublicFetchOpts{
+			opts := config.PublicFetchOpts{
 				Client: httpClient,
 				URL:    u,
 			}
 			klog.V(2).Infof("Provisioning from %q...", u)
-			logs, err := FetchPublicConfig(ctx, opts)
+			logs, err := config.FetchPublicConfig(ctx, opts)
 			if err != nil {
 				klog.Warningf("Failed to fetch public witness network config from %q: %v", u, err)
 				continue
