+++ persistence/sqlite/sql.go
@@ -24,7 +24,7 @@ import (
 	"github.com/transparency-dev/formats/log"
 	"github.com/transparency-dev/formats/note"
 	"github.com/transparency-dev/witness/witness"
-	"github.com/transparency-dev/witness/omniwitness"
+	"github.com/transparency-dev/witness/config"
 	"k8s.io/klog/v2"
 
 	"modernc.org/sqlite"
@@ -85,7 +85,7 @@ func (p *Persistence) createTablesIfNotExist(ctx context.Context) error {
 	return nil
 }
 
-func (p *Persistence) AddLogs(ctx context.Context, lc []omniwitness.Log) error {
+func (p *Persistence) AddLogs(ctx context.Context, lc []config.Log) error {
 	for _, l := range lc {
 		// Note that it's a deliberate choice here to use Insert so as to guarantee that we will not
 		// update the stored config for a given log. There may be reasons to revisit this in the future,
@@ -100,11 +100,11 @@ func (p *Persistence) AddLogs(ctx context.Context, lc []omniwitness.Log) error {
 	return nil
 }
 
-func (p *Persistence) Logs(ctx context.Context) iter.Seq2[omniwitness.Log, error] {
-	return func(yield func(omniwitness.Log, error) bool) {
+func (p *Persistence) Logs(ctx context.Context) iter.Seq2[config.Log, error] {
+	return func(yield func(config.Log, error) bool) {
 		rows, err := p.db.QueryContext(ctx, "SELECT origin, vkey, contact, disabled FROM logs")
 		if err != nil {
-			if !yield(omniwitness.Log{}, fmt.Errorf("failed to select from logs: %v", err)) {
+			if !yield(config.Log{}, fmt.Errorf("failed to select from logs: %v", err)) {
 				return
 			}
 		}
@@ -112,10 +112,10 @@ func (p *Persistence) Logs(ctx context.Context) iter.Seq2[omniwitness.Log, error
 			_ = rows.Close()
 		}()
 		for rows.Next() {
-			c := omniwitness.Log{}
+			c := config.Log{}
 			disabled := false
 			if err := rows.Scan(&c.Origin, &c.VKey, &c.Contact, &disabled); err != nil {
-				if !yield(omniwitness.Log{}, fmt.Errorf("failed to scan columns: %v", err)) {
+				if !yield(config.Log{}, fmt.Errorf("failed to scan columns: %v", err)) {
 					return
 				}
 			}
@@ -125,7 +125,7 @@ func (p *Persistence) Logs(ctx context.Context) iter.Seq2[omniwitness.Log, error
 			}
 			c.Verifier, err = note.NewVerifier(c.VKey)
 			if err != nil {
-				if !yield(omniwitness.Log{}, fmt.Errorf("failed to create verifier: %v", err)) {
+				if !yield(config.Log{}, fmt.Errorf("failed to create verifier: %v", err)) {
 					return
 				}
 			}
@@ -136,16 +136,16 @@ func (p *Persistence) Logs(ctx context.Context) iter.Seq2[omniwitness.Log, error
 	}
 }
 
-func (p *Persistence) Log(ctx context.Context, origin string) (omniwitness.Log, bool, error) {
+func (p *Persistence) Log(ctx context.Context, origin string) (config.Log, bool, error) {
 	logID := log.ID(origin)
 	row := p.db.QueryRowContext(ctx, "SELECT origin, vkey, contact, disabled FROM logs WHERE logID = ?", logID)
 	if row.Err() != nil {
-		return omniwitness.Log{}, false, fmt.Errorf("failed to select from logs: %v", row.Err())
+		return config.Log{}, false, fmt.Errorf("failed to select from logs: %v", row.Err())
 	}
-	c := omniwitness.Log{}
+	c := config.Log{}
 	disabled := false
 	if err := row.Scan(&c.Origin, &c.VKey, &c.Contact, &disabled); err != nil {
-		return omniwitness.Log{}, false, fmt.Errorf("failed to scan columns: %v", err)
+		return config.Log{}, false, fmt.Errorf("failed to scan columns: %v", err)
 	}
 	if disabled {
 		klog.V(1).Infof("Ignoring disabled log %q", c.Origin)
@@ -154,7 +154,7 @@ func (p *Persistence) Log(ctx context.Context, origin string) (omniwitness.Log, 
 	var err error
 	c.Verifier, err = note.NewVerifier(c.VKey)
 	if err != nil {
-		return omniwitness.Log{}, false, fmt.Errorf("failed to create verifier: %v", err)
+		return config.Log{}, false, fmt.Errorf("failed to create verifier: %v", err)
 	}
 	return c, true, nil
 }
